moved donut data processing function to chart helpers.
